+++ src/main/kotlin/land/tbp/y2019/intcode/computer/Memory.kt
@@ -7,11 +7,11 @@ class Memory(ints: List<Int>) {
         ints[position] = value
     }
 
-    fun read(position: Int): Int {
+    fun read(position: Int): Long {
         return ints[position]
     }
 
-    fun contents(): List<Int> {
+    fun contents(): List<Long> {
         return ints.toList()
     }
